// src/app/interfaces/toShow.ts
export interface SongToShow {
  id_song: number
  img_src: string
  name_song: string
  price_song: number
}
export interface FoodToShow {
  id_food: number
  img_food: string
  name_food: string
  price_food: number
}
export interface LocationShow {
  id_location: number
  img_location: string
  name_location: string
  price_location: number
}
export interface ThemeShow {
  id_theme: number
  img_theme: string
  name_theme: string
  price_theme: number
}